103: fix кодревью - драйверы (SSD1306/ST7789/ILI9225), шрифты, BMP-загрузчики, easing, transformation, resource

+++ Display/uTFT2/Resource/resource.h
@@ -8,7 +8,7 @@ extern void setResStartAdress(uint32_t i);
 //Получить начало ресурсов
 extern uint8_t * getResAdressFontID(uint32_t id);
 
-//Получить дескриптов картинки из ресурсов 16 и 32 бит
-extern Bitmap getResBitmapID(int8_t id);
+//Получить дескриптор картинки из ресурсов 16 и 32 бит
+extern Bitmap getResBitmapID(uint32_t id);
 
 #endif /* RESOURCE_RESOURCE_H_ */
